+++ koku/api/tags/aws/aws_tag_query_handler.py
@@ -15,12 +15,6 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 """OCP Tag Query Handling."""
-import copy
-
-from django.db.models import Count
-from tenant_schemas.utils import tenant_context
-
-from api.functions import JSONBObjectKeys
 from api.tags.queries import TagQueryHandler
 from reporting.models import AWSCostEntryLineItemDailySummary
 
@@ -39,78 +33,4 @@ class AWSTagQueryHandler(TagQueryHandler):
             kwargs    (Dict): A dictionary for internal query alteration based on path
         """
         super().__init__(query_parameters, url_data,
-                         tenant, **kwargs)
-
-    def _format_query_response(self):
-        """Format the query response with data.
-
-        Returns:
-            (Dict): Dictionary response of query params, data, and total
-
-        """
-        output = copy.deepcopy(self.query_parameters)
-        output['data'] = self.query_data
-
-        return output
-
-    def get_tag_keys(self, filters=True):
-        """Get a list of tag keys to validate filters."""
-        with tenant_context(self.tenant):
-            tag_keys = AWSCostEntryLineItemDailySummary.objects
-            if filters is True:
-                tag_keys = tag_keys.filter(self.query_filter)
-
-            tag_keys = tag_keys.annotate(tag_keys=JSONBObjectKeys('tags'))\
-                .values('tag_keys')\
-                .annotate(tag_count=Count('tag_keys'))\
-                .all()
-            tag_keys = [tag.get('tag_keys') for tag in tag_keys]
-
-        return tag_keys
-
-    def get_tags(self):
-        """Get a list of tags and values to validate filters."""
-        def get_dictionary_for_key(merged_data, key):
-            for di in merged_data:
-                if key in di.get('key'):
-                    return di
-            return None
-
-        with tenant_context(self.tenant):
-            tag_keys = AWSCostEntryLineItemDailySummary.objects\
-                .filter(self.query_filter)\
-                .values('tags')\
-                .all()
-            tag_keys = [tag.get('tags') for tag in tag_keys]
-
-            merged_data = []
-            for item in tag_keys:
-                for key, value in item.items():
-                    key_dict = get_dictionary_for_key(merged_data, key)
-                    if not key_dict:
-                        new_dict = {}
-                        new_dict['key'] = key
-                        new_dict['values'] = [value]
-                        merged_data.append(new_dict)
-                    else:
-                        if value not in key_dict.get('values'):
-                            key_dict['values'].append(value)
-                            key_dict['values'].sort()
-        return merged_data
-
-    def execute_query(self):
-        """Execute query and return provided data.
-
-        Returns:
-            (Dict): Dictionary response of query params and data
-
-        """
-        if self.query_parameters.get('key_only'):
-            tag_keys = self.get_tag_keys()
-            query_data = sorted(tag_keys, reverse=self.order_direction == 'desc')
-        else:
-            tags = self.get_tags()
-            query_data = sorted(tags, key=lambda k: k['key'], reverse=self.order_direction == 'desc')
-
-        self.query_data = query_data
-        return self._format_query_response()
+                         tenant, AWSCostEntryLineItemDailySummary, **kwargs)
